requirements.txt
# STRYDER AI — Python Dependencies
# ==================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
joblib>=1.3.0
groq>=0.4.0
supabase>=2.0.0
osmnx>=1.7.0
networkx>=3.2.0
geopy>=2.4.0
folium>=0.15.0
httpx>=0.25.0
pydantic>=2.5.0
sse-starlette>=1.8.0
gunicorn>=21.2.0
